add sign up and sign in methods by jwt token

+++ internal/models/film.go
@@ -1,7 +1,7 @@
 package models
 
 type Film struct {
-	Id          int32 `gorm:"primaryKey"`
+	Id          int64 `gorm:"primaryKey"`
 	Title       string
 	Description string
 	Genre       string
